Moved main classes
JavaPlatform should be in mocuishle2 maybe but cache depends on it.
